+++ quill/test/integration_tests/ArithmeticArrayTypesLoggingTest.cpp
@@ -130,6 +130,10 @@ TEST_CASE("arithmetic_array_types_logging")
     LOG_INFO(logger, "tc {}", tc);
   }
 
+  int v = 111;
+  int* ptr_test = &v;
+  LOG_INFO(logger, "pointer [{}]", static_cast<void const*>(ptr_test));
+
   logger->flush_log();
   Frontend::remove_logger(logger);
 
@@ -162,6 +166,11 @@ TEST_CASE("arithmetic_array_types_logging")
 
   REQUIRE(quill::testing::file_contains(
     file_contents, std::string{"LOG_INFO      " + logger_name + "       tc Test5"}));
+
+  std::string expected_ptr_value_str{"LOG_INFO      " + logger_name + "       pointer ["};
+  expected_ptr_value_str += fmtquill::format("{}", fmtquill::ptr(ptr_test));
+  expected_ptr_value_str += "]";
+  REQUIRE(quill::testing::file_contains(file_contents, expected_ptr_value_str));
 
   testing::remove_file(filename);
 }